+++ entrypoints/popup/main.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import './style.css';
+
 import Popup from './popup';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(